graphqlのurlにapiを入れて、interfaceにresponseをつける
