+++ app/frontend/js/modules/ruka/graphql.ts
@@ -3,19 +3,19 @@ import { axios } from '../../lib/axios';
 import { ResponseInterface } from '../../lib/interface/response';
 import { AxiosResponse } from 'axios';
 import { GirlModel } from '../../lib/interface/model';
-import { buildGraphqlUrl } from '../../lib/utils/url_builder';
+import { buildGraphqlApiUrl } from '../../lib/utils/url_builder';
 
 type queryObj = {
   operation: string;
   fields: string[];
 };
 
-export interface GirlsInterface extends ResponseInterface {
+export interface ReponseGirlsInterface extends ResponseInterface {
   data: {
     girls: GirlModel[];
   };
 }
 
-export const graphqlQuery = async (args: queryObj): Promise<AxiosResponse<GirlsInterface>> => {
-  return axios().post(buildGraphqlUrl(), query(args));
+export const graphqlQuery = async (args: queryObj): Promise<AxiosResponse<ReponseGirlsInterface>> => {
+  return axios().post(buildGraphqlApiUrl(), query(args));
 };
