+++ config/routes.rb
@@ -2,10 +2,10 @@
 
 Rails.application.routes.draw do
   mount GraphiQL::Rails::Engine, at: '/graphiql', graphql_path: '/graphql' if Rails.env.development?
-  post '/graphql', to: 'graphql#execute'
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 
   namespace :api do
+    post '/graphql', to: 'graphql#execute'
     resources :girls, only: %i[index show]
   end
